+++ libft/ft_substr.c
@@ -4,7 +4,7 @@ char	*ft_substr(char const *s, unsigned int start, size_t len){
 	char	*substr;
 	int		i;
 
-	if (s && start < ft_strlen((char *) s)){
+	if (s && start < (unsigned int) ft_strlen((char *) s)){
 		if (!len)
 			return (ft_strdup(""));
 		substr = (char *) malloc ((len + 1) * sizeof(char));
